requirements.txt
absl-py==0.15.0
actionlib==1.12.1
angles==1.9.12
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
backports.functools-lru-cache==1.6.4
bleach==1.5.0
bondpy==1.8.5
bzr==2.8.0.dev1
cachetools==3.1.1
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
catkin-tools==0.6.1
cbor==1.0.0
certifi==2021.10.8
chardet==4.0.0
click==6.7
cloudpickle==1.3.0
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
controller-manager==0.18.4
controller-manager-msgs==0.18.4
cryptography==2.1.4
cv-bridge==1.13.0
cycler==0.10.0
Cython==0.26.1
decorator==4.4.2
defusedxml==0.5.0
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
docutils==0.14
dynamic-reconfigure==1.6.3
empy==3.3.2
enum34==1.1.10
future==0.18.2
futures==3.3.0
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
google-auth==1.35.0
google-auth-oauthlib==0.4.1
greenlet==0.4.12
grpcio==1.41.1
html5lib==0.9999999
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
image-geometry==1.13.0
incremental==16.10.1
interactive-markers==1.11.5
ipaddress==1.0.17
joint-state-publisher==1.12.15
kdl-parser-py==1.13.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
laser-geometry==1.6.7
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lz4==0.10.1
Markdown==3.1.1
matplotlib==2.2.5
mavros==1.9.0
mercurial==4.5.3
message-filters==1.14.11
mpi4py==2.0.0
msgpack==0.5.6
neovim==0.2.0
netifaces==0.10.4
networkx==2.2
nose==1.3.7
numpy==1.16.6
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
osrf-pycommon==0.2.1
PAM==0.4.2
paramiko==2.0.0
Pillow==6.2.2
protobuf==3.17.0
psutil==5.4.2
py-ubjson==0.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.4.7
pydot==1.2.3
Pygments==2.2.0
pygobject==3.26.1
PyNaCl==1.1.2
PyOpenGL==3.1.0
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.2
python-gnupg==0.4.1
python-qt-binding==0.4.4
python-snappy==0.5
PyTrie==0.2
pytz==2021.3
PyWavelets==1.0.3
pyxdg==0.25
PyYAML==3.12
qrcode==5.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.26.0
requests-oauthlib==1.3.0
resource-retriever==1.12.6
roman==2.0.0
rosbag==1.14.11
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosdep==0.21.0
rosdep-modules==0.21.0
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.14.11
rosinstall==0.7.8
rosinstall-generator==0.1.22
roslaunch==1.14.11
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.11
rosmake==1.14.9
rosmaster==1.14.11
rosmsg==1.14.11
rosnode==1.14.11
rosparam==1.14.11
rospkg==1.3.0
rospkg-modules==1.3.0
rospy==1.14.11
rosservice==1.14.11
rostest==1.14.11
rostopic==1.14.11
rosunit==1.14.9
roswtf==1.14.11
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.10
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.7.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rsa==4.5
rviz==1.13.18
scikit-image==0.14.5
scipy==1.2.3
SecretStorage==2.3.1
sensor-msgs==1.12.8
service-identity==16.0.0
simplejson==3.13.2
six==1.16.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
subprocess32==3.5.4
tensorboard==2.1.0
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.11
torch==1.4.0
trollius==2.0.1
Twisted==17.9.0
txaio==2.8.1
u-msgpack-python==2.1
urdfdom-py==0.4.5
urllib3==1.26.7
uuv-assistants==0.6.13
uuv-auv-control-allocator==0.6.13
uuv-control-cascaded-pid==0.6.13
uuv-thruster-manager==0.6.13
uuv-trajectory-control==0.6.13
vcstools==0.1.42
wadllib==1.3.2
Werkzeug==1.0.1
wsaccel==0.6.2
wstool==0.1.17
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xacro==1.13.12
zope.interface==4.3.2
